refactor(drift): 抽象 drift 入口
